Move to class-based message handlers.
This new form of message handler allows wrapping the handler method
in a universal wrapper function, which ensures certain mandatory
fields are present, and wraps the payload in the mandatory reply
format.

Since otherwise every handler would have to do this in the exact
same way, this greatly reduces future code duplication.